requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
google-generativeai